revert to input validation errors funneled to endpoint error catch
