+++ backend/helpers/handleError.js
@@ -4,8 +4,6 @@ Server Error Handling Helper | Capstone App (Pursuit Volunteer Mgr)
 */
 
 
-const errorResponse = require('./errorResponse');
-
 /*
 USAGE
 to instantiate custom error, format: throw new Error(`<error code>__error: <msg>`);
@@ -13,38 +11,49 @@ example: throw new Error(`404__error: volunteer ${volunteerId} not found`);
 */
 
 
+const errorResponse = (res, code, errorMsg) => {
+  console.log(errorMsg);   // log for developer
+  res.status(code);
+  res.json({
+      status: "fail",
+      message: errorMsg,
+      payload: null
+  });
+}
+
+
 const handleError = (err, req, res, next) => {
   // per documentation, checking for present headers and if so, passing
   if (res.headersSent) {
-    console.log("err: res headers already exist. passing error to express");
+    console.log("err: res headers already exist. passing error to express");  // msg to developer
     return next(err);
-  }
-
-  // catch and parse error.messages for relevant http code
-  let [ code, msg = code ] = err.message.split('__');
-
-  // if "__" wasn't found, error was not thrown from a customized instance
-  if (msg === code) {
-    errorResponse(res, 500, `(back) error: ${msg}`);
   } else {
 
-    // specific-case errors
+    // specific-case (database response) errors
     if (err.message === "No data returned from the query.") {
-      errorResponse(res, 404, `(front) error: specific record does not exist`);
-    }
-    if (err.message.includes("violates unique constraint")) {
-      errorResponse(res, 403, `(front) error: at least one unique datapoint of record already exists`);
+      err.message = `404__error: specific record does not exist`;
+    } else if (err.message.includes("violates unique constraint")) {
+      err.message = `403__error: at least one unique datapoint of record already exists`;
     }
 
-    // custom error handling
-    res.status(parseInt(code));
-    const codeClass = parseInt(code).toString()[0];
-    if (codeClass === '4') {
-      errorLocation = "(front)";
+    // catch and parse error.messages for relevant http code
+    let [ code, msg = code ] = err.message.split('__');
+
+    // if "__" wasn't found, error was not thrown from a customized instance
+    if (msg === code) {
+      errorResponse(res, 500, `(back) error: ${msg}`);
     } else {
-      errorLocation = "(back)";
+
+      // create and send response
+      res.status(parseInt(code));
+      const codeClass = parseInt(code).toString()[0];
+      if (codeClass === '4') {
+        errorLocation = "(front)";
+      } else {
+        errorLocation = "(back)";
+      }
+      errorResponse(res, code, `${errorLocation} ${msg}`);
     }
-    errorResponse(res, code, `${errorLocation} error: ${msg}`);
   }
 };
